+++ include/hfts_grasp_planner/placement/mp/MultiGraspMP.h
@@ -11,9 +11,9 @@ namespace mp {
     public:
         struct Grasp {
             unsigned int id;
-            OpenRAVE::geometry::RaveVector<float> pos;
-            OpenRAVE::geometry::RaveVector<float> quat;
-            std::vector<float> gripper_values;
+            OpenRAVE::geometry::RaveVector<double> pos;
+            OpenRAVE::geometry::RaveVector<double> quat;
+            std::vector<double> gripper_values;
 
             std::string print() const
             {
@@ -22,7 +22,7 @@ namespace mp {
                    << "x=" << pos.x << " y=" << pos.y << " z=" << pos.z
                    << " qx=" << quat.x << " qy=" << quat.y << " qz=" << quat.z << " qw=" << quat.w;
                 ss << " q=[";
-                for (const float& v : gripper_values) {
+                for (const double& v : gripper_values) {
                     ss << v << ", ";
                 }
                 ss << "]";
@@ -32,23 +32,23 @@ namespace mp {
         struct Goal {
             unsigned int id;
             unsigned int grasp_id;
-            std::vector<float> config;
+            std::vector<double> config;
             std::string print() const
             {
                 std::stringstream ss;
                 ss << "id=" << id << " "
                    << "grasp_id=" << grasp_id;
                 ss << " q=[";
-                for (const float& v : config) {
+                for (const double& v : config) {
                     ss << v << ", ";
                 }
                 ss << "]";
                 return ss.str();
             }
         };
-        typedef std::vector<std::vector<float>> WaypointPath;
+        typedef std::vector<std::vector<double>> WaypointPath;
         virtual ~MultiGraspMP() = 0;
-        virtual void plan(std::vector<std::pair<unsigned int, WaypointPath>>& new_paths, float time_limit = 0.0f) = 0;
+        virtual void plan(std::vector<std::pair<unsigned int, WaypointPath>>& new_paths, double time_limit = 0.0f) = 0;
         virtual void addGrasp(const Grasp& grasp) = 0;
         virtual void addGoal(const Goal& goal) = 0;
         virtual void removeGoals(const std::vector<unsigned int>& goal_ids) = 0;
